promptlab_backend/requirements.txt


APScheduler==3.11.0
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
blis==1.3.0
breadability==0.1.20
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5

contourpy==1.3.0
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
datasets==4.0.0
deep-translator==1.11.4
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
docopt==0.6.2
es_core_news_md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl#sha256=478b8bb3f3e8eb149192f7d80e7cd64f990b7e9bccbc5329df41a57e86326be2
et_xmlfile==2.0.0
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
graphviz==0.21
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
imbalanced-learn
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lightgbm==4.6.0
llama_cpp_python==0.3.16
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.3
nltk==3.9.1
numpy==2.1.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
plotly==6.2.0
postgrest==1.1.1
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sklearn-compat
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==5.0.0
SQLAlchemy==2.0.42
sqlparse==0.5.3
srsly==2.5.1
storage3==0.12.0
StrEnum==0.4.15
sumy==0.11.0
supabase==2.17.0
supafunc==0.10.1
sympy==1.13.3
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
rq==1.15.1
tqdm==4.67.1
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.3
xxhash==3.5.0
yarl==1.20.1
cryptography==45.0.6
rembg==2.0.67
onnxruntime== 1.22.1
ultralytics==8.3.196
pytesseract==0.3.13
#redis== 6.4.0
yellowbrick==1.5